chore: split seed_db.py into per-scenario modules
scripts/seed_db.py was 1599 lines, over house-lint's 800-line HSL102
threshold. The file was a sequence of independent, same-shaped scenario
generators plus a set of shared helpers, so the split needed no design
decision — each scenario moved to its own module as-is.

- scripts/seed_scenarios/base.py holds the shared surface: SeedIntegrityError,
  SeedContext, the INSERT-SQL builders, the seed constants, and the seed_*
  building blocks.
- One module per scenario (empty, healthy, degraded, error, large_volume,
  lifecycle, adversarial), with the registry in seed_scenarios/__init__.py.
- seed_db.py keeps the CLI, the transaction/atomic-swap machinery, and the
  post-seed integrity checks, and re-exports SCENARIOS so seed_db.SCENARIOS
  stays the same dict object.

Helpers that now cross module boundaries lost their underscore prefix
(_seed_listener -> seed_listener, etc.); ones still local to base.py kept it.
The per-block dup-ignore pairs became one dup-ignore-file marker per affected
scenario module — after the split PMD's fragment boundaries spilled past the
old end markers and raised a new cross-file cluster.

No behavior change: all 7 scenarios produce logically identical databases
(SQLite dumps compared before/after, masking only the app_manifests wall-clock
columns that already vary run-to-run), and every moved definition is textually
identical to the pre-split file apart from the renames.
